Changed count() operations to return long instead of int because this is how the new Hibernate 4.1 API works.

+++ features/access-point-monitor/src/main/java/org/opennms/netmgt/accesspointmonitor/rest/AccessPointRestService.java
@@ -118,7 +118,7 @@ public class AccessPointRestService {
     public String getCount() {
         readLock();
         try {
-            return Integer.toString(m_accessPointDao.countAll());
+            return Long.toString(m_accessPointDao.countAll());
         } finally {
             readUnlock();
         }
