+++ opennms-webapp/src/main/java/org/opennms/web/rest/OutageRestService.java
@@ -117,7 +117,7 @@ public class OutageRestService extends OnmsRestService {
     public String getCount() {
         readLock();
         try {
-            return Integer.toString(m_outageDao.countAll());
+            return Long.toString(m_outageDao.countAll());
         } finally {
             readUnlock();
         }
